convert array to float

+++ cali_beta_pic.py
@@ -84,7 +84,7 @@ def reduce_images(out_path, prefix_filenames):
             fd = (fd - master_bias) / master_flat
 
             # Convert to np.float32
-            fd = fd.astype(np.float32)
+            fd = fd.astype(float)
 
             reduced_data.append(fd)  # Append the reduced image to the list
             reduced_header_info.append(hdr)
